kernel: x86: tighten shift and return-value handling
Address scanner findings with no intended behavioral changes:
- force LOAPIC logical-ID shift operands to uint32_t in intc_loapic,
- handle k_mutex_unlock()/k_mutex_lock() return values in k_condvar_wait,
- make intentional discard of z_add_thread_timeout() explicit in
  thread_schedule_new().

+++ drivers/interrupt_controller/intc_loapic.c
@@ -101,7 +101,8 @@ void z_loapic_enable(unsigned char cpu_number)
 	 *
 	 * in X2APIC, LDR is read-only.
 	 */
-	x86_write_xapic(LOAPIC_LDR, 1U << (cpu_number + 24));
+	x86_write_xapic(LOAPIC_LDR,
+			((uint32_t)1U) << (((uint32_t)cpu_number) + 24U));
 #endif
 
 	/*
